fix: address 7 low-severity polish issues
1. Shift amount truncation (value.rs): use u32::try_from instead of
   `as u32` cast to avoid silent truncation for i64/i128/u64/u128 shifts
2. VarId overflow (eval.rs): add debug_assert before increment
3. SourceId overflow (eval.rs): add debug_assert before increment
4. show_changes isize cast (main.rs): use i64 for delta subtraction
5. Qualified call path (ascent-ir): join all path segments with `::`
6. unwrap on syn::parse2 (ascent-ir): replace with descriptive expect
7. coerce_to_col_type (eval.rs): add doc comment explaining fallback behavior

+++ src/main.rs
@@ -324,7 +324,7 @@ fn show_changes(engine: &mut Engine, prev_counts: &mut HashMap<String, usize>) {
         prev_counts.insert(name.clone(), count);
 
         if count != prev {
-            let delta = count as isize - prev as isize;
+            let delta = count as i64 - prev as i64;
             let sign = if delta > 0 { "+" } else { "" };
             eprintln!(
                 "  {name}: {count} tuple{} ({sign}{delta})",
